requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
typing-extensions==4.11.0
Pillow==10.4.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
